RSpec support: change before hook to around hook (#17)
In order to allow end-users to use `around` hooks in RSpec setup, we need to run our library setup in an `around` hook. Otherwise, our `before` hook will not have run when the end-user `around` hook fires, and that means that any WebValve-dependent behavior in that `around` hook will fail, surprisingly.