default services to enabled with webvalve_enabled=1 in production
this changes the configuration of webvalve_enabled=1 for production-like
environments to better align with our intuitions.

So what if the defaults look like this:

|  | Dev/Test  | Prod-like |
| -----|-------- | ------------- |
| `WEBVALVE_ENABLED` | true | false  |
| `[SERVICE]_ENABLED` | false  | true |

That way even if you set `WEBVALVE_ENABLED` in stage/prod/demo, you have
to opt services _in_ to being faked instead of opting them out.