change WebValve.register to accept a class name instead of class (#35)
* change WebValve.register to accept a class name instead of class

rails 6 autoloading emits a deprecation warning for class loading during
app initialization. webvalve's register call is a culprit here, but it
doesn't have to be. this changeset makes it so that webvalve accepts a
stringified class name instead of a class and makes that all work
properly with the internals.

* update and correct changelog

* fix CI, update version and fix changelog

* add 2.6.1